requirements.txt
Django==5.0
#static file uchun
#hosting uchun 
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
pillow==10.4.0
#back end uchun
djangorestframework==3.15.1
django-cors-headers==4.3.1